P2b: extract afxdp/tx/rings.rs (XSK kernel-ring discipline, 6 fns ~280 LOC)
Extracts the XSK kernel-ring helpers from tx/mod.rs into sibling
tx/rings.rs:
- reap_tx_completions (XSK completion ring drain) — pub(in crate::afxdp).
- drain_pending_fill (XSK fill ring submit) — bumped from pub(super)
  to pub(in crate::afxdp); facade re-export pub(super) from tx/mod.rs
  preserves frame_tx.rs / afxdp.rs sibling visibility.
- maybe_wake_rx (kernel RX wake) — same visibility bump pattern.
- maybe_wake_tx (kernel TX wake) — pub(in crate::afxdp).
- recycle_completed_tx_offset — file-private (sole caller is moving).
- apply_prepared_recycle — pub(super) in rings.rs + #[cfg(test)] use
  in tx/mod.rs (test pin at tx/mod.rs:3329 keeps working through
  super::* resolution).

tx/mod.rs adds 'pub(super) mod rings;' + the mixed-visibility re-export
block. Stats re-exports split: stamp_submits stays always-on (production
callers in cos/queue_service.rs); record_kick_latency and
record_tx_completions_with_stamp are now #[cfg(test)]-gated since
umem.rs::tests is the only consumer (production callers inside
tx/rings.rs import from super::stats::* directly).

Constants (FILL_BATCH_SIZE, FILL_WAKE_SAFETY_INTERVAL_NS, RX_WAKE_*,
TX_WAKE_MIN_INTERVAL_NS, XskBindMode) are reached from afxdp.rs via
crate::afxdp::* — no move needed; same visibility-default pattern as
P2a's UMEM_FRAME_SHIFT.

cargo build --bins clean (88 warnings, all pre-existing dead-code; 0
unused-imports new). cargo test --bins 865/0/2 (rolling baseline).

Plan: docs/pr/p2b-tx-rings/plan.md (v3.1, addressing Codex r1+r2+r3
findings).
